Do banana checks on the client side; anonymously log requests.
Banana longpolls which is a killer for serverless.  Now we start
the request, send the callID to the client who can longpoll for the
result.

On server, we also (anonymously) log the request and startTime.
Client updates this on getting the result from pollPoll.  Then
we can compute the diff and log the total execution time in
the database, per model_id.